refactor: modify the directory view as the map
