+++ admin/map/src/components/mapClient.jsx
@@ -9,6 +9,7 @@ import { MapContainer, Marker, Popup, TileLayer } from 'react-leaflet'
 
 import 'leaflet/dist/leaflet.css'
 import '@changey/react-leaflet-markercluster/dist/styles.min.css'
+import { schemaContent } from '../utils/schemaContent'
 
 delete L.Icon.Default.prototype._getIconUrl
 
@@ -37,121 +38,7 @@ const markerClicked = async (profile, apiUrl, linkType) => {
       `Error getting post, please contact the administrator, error: ${responseData}`
     )
   }
-  let content = ''
-  const schema = responseData?.profile_data?.linked_schemas[0]
-  const imageUrl = responseData?.profile_data?.image
-  switch (schema) {
-    case 'organizations_schema-v1.0.0':
-      content = addImageToContent(content, imageUrl)
-      content = addTitleToContent(content, responseData?.profile_data?.name)
-      content = addDescriptionToContent(
-        content,
-        responseData?.profile_data?.description
-      )
-      content = addUrlToContent(
-        content,
-        responseData?.profile_data?.primary_url,
-        responseData.post_url,
-        linkType
-      )
-      break
-    case 'people_schema-v0.1.0':
-      content = addImageToContent(content, imageUrl)
-      content = addTitleToContent(
-        content,
-        responseData?.profile_data?.full_name
-      )
-      content = addDescriptionToContent(
-        content,
-        responseData?.profile_data?.description
-      )
-      content = addUrlToContent(
-        content,
-        responseData?.profile_data?.primary_url,
-        responseData.post_url,
-        linkType
-      )
-      break
-    case 'offers_wants_schema-v0.1.0':
-      content = addImageToContent(content, imageUrl)
-      content = addTitleToContent(content, responseData?.profile_data?.title)
-      content = addTextToContent(
-        content,
-        responseData?.profile_data?.exchange_type
-      )
-      content = addTextToContent(
-        content,
-        responseData?.profile_data?.transaction_type
-      )
-      content = addDescriptionToContent(
-        content,
-        responseData?.profile_data?.description
-      )
-      content = addUrlToContent(
-        content,
-        responseData?.profile_data?.details_url,
-        responseData.post_url,
-        linkType
-      )
-      break
-  }
-
-  return content
-}
-
-function addImageToContent(content, imageUrl) {
-  if (imageUrl) {
-    content += `<img src='${imageUrl}' style='height: 50px; width: auto' id='profile_image' alt="" />`
-
-    const img = new Image()
-    img.src = imageUrl
-    img.onerror = () => {
-      const profileImage = document.getElementById('profile_image')
-      if (profileImage) {
-        profileImage.style.display = 'none'
-      }
-    }
-  }
-  return content
-}
-
-function addTitleToContent(content, title) {
-  if (title) {
-    content += `<p><strong>${title}</strong></p>`
-  }
-  return content
-}
-
-function addTextToContent(content, text) {
-  if (text) {
-    content += `<p>${text}</p>`
-  }
-  return content
-}
-
-function addDescriptionToContent(content, description) {
-  if (description) {
-    content += `<p>${limitString(description, 200)}</p>`
-  }
-  return content
-}
-
-function addUrlToContent(content, primaryUrl, postUrl, linkType) {
-  if (primaryUrl && linkType === 'primary') {
-    content += `<p><a target='_blank' rel='noreferrer' href='${primaryUrl}'>${primaryUrl}</a></p>`
-  }
-  if (postUrl && linkType === 'wp') {
-    content += `<p><a href='${postUrl}'>More...</a></p>`
-  }
-  return content
-}
-
-function limitString(inputString, maxLength) {
-  if (inputString.length <= maxLength) {
-    return inputString
-  } else {
-    return inputString.substring(0, maxLength) + '...'
-  }
+  return schemaContent(responseData, linkType)
 }
 
 export default function MapClient({
